semantic/naming updates to api router and controller; add comments to prepare replacing buttons with links; move POST form for adding songs within forEach; uncomment app.use 404 in server.js; comment out api (songs) model; update property fields for songs array in playlistSchema
